requirements.txt
attrs==22.1.0
autopep8==2.0.0
aws-cdk-lib==2.8.0
cattrs==22.2.0
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
constructs==10.1.180
distlib==0.3.6
exceptiongroup==1.0.4
filelock==3.8.2
identify==2.5.9
idna==3.4
iniconfig==1.1.1
jsii==1.72.0
nodeenv==1.7.0
packaging==21.3
platformdirs==2.5.4
pluggy==1.0.0
pre-commit==2.20.0
publication==0.0.3
pycodestyle==2.10.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
toml==0.10.2
tomli==2.0.1
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.17.0
